# docs/swagger.yaml
basePath: /
definitions:
  controllers.RunHourlySimulationRequest:
    properties:
      simulated_at:
        type: string
    type: object
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.CreateUserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      expires_in:
        type: integer
      fincode_customer_id:
        type: string
      fincode_subscription_id:
        type: string
      id:
        type: string
      subsc:
        type: boolean
      token:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    type: object
  dto.LoginResponse:
    properties:
      expires_in:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      fincode_customer_id:
        type: string
      fincode_subscription_id:
        type: string
      id:
        type: string
      subsc:
        type: boolean
    type: object
...
  usecases.FindByTodayReportOutput:
    properties:
      createdAt:
        type: string
      gossip:
        type: string
      hourSlot:
        type: integer
      id:
        type: string
      petID:
        type: string
    type: object
  usecases.RunHourlyPetSimulationOutput:
    properties:
      processed:
        type: integer
      results:
        items:
          $ref: '#/definitions/usecases.RunHourlyPetSimulationPetResult'
        type: array
      simulated_at:
        type: string
      skipped:
        type: integer
      total_pets:
        type: integer
    type: object
  usecases.RunHourlyPetSimulationPetResult:
    properties:
      ambient_event:
        type: string
      move_probability:
        type: number
      moved:
        type: boolean
      next_group_master_id:
        type: integer
      pet_id:
        type: string
      previous_group_master_id:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: PETYO-YO backend API documentation
  title: PETYO-YO API
  version: "1.0"
paths:
...
  /reports/{pet_id}:
    get:
      description: 指定したペットIDの当日分レポートを取得します。
      parameters:
      - description: ペットID
        in: path
        name: pet_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 取得成功
          schema:
            items:
              $ref: '#/definitions/usecases.FindByTodayReportOutput'
            type: array
        "400":
          description: ペットID不正
          schema:
            type: string
        "500":
          description: サーバーエラー
          schema:
            type: string
      summary: 当日レポート取得
      tags:
      - reports
  /simulations/hourly:
    post:
      consumes:
      - application/json
      description: 全アクティブペットを対象に、指定時刻または現在時刻の時間単位シミュレーションを実行します。
      parameters:
      - description: シミュレーション条件
        in: body
        name: request
        schema:
          $ref: '#/definitions/controllers.RunHourlySimulationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 実行成功
          schema:
            $ref: '#/definitions/usecases.RunHourlyPetSimulationOutput'
        "400":
          description: リクエストまたはマスターデータ不正
          schema:
            type: string
        "405":
          description: 許可されていないメソッド
          schema:
            type: string
        "500":
          description: サーバーエラー
          schema:
            type: string
      summary: 時間単位シミュレーション実行
      tags:
      - simulations
...
  /users:
    post:
      consumes:
      - application/json
      description: メールアドレスとパスワードを使用して新規ユーザーを作成し、認証トークンを返します。
      parameters:
      - description: ユーザー情報
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 登録成功
          schema:
            $ref: '#/definitions/dto.CreateUserResponse'
        "400":
          description: リクエスト不正
          schema:
            type: string
        "405":
          description: 許可されていないメソッド
          schema:
            type: string
        "500":
          description: サーバーエラー
          schema:
            type: string
      summary: ユーザー登録
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: メールアドレスとパスワード、または Authorization ヘッダーのトークンを使用して認証します。
      parameters:
      - description: ログイン情報（トークン認証時は省略可能）
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      - description: Bearer トークン
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ログイン成功
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: 認証情報不足またはリクエスト不正
          schema:
            type: string
        "401":
          description: 認証失敗
          schema:
            type: string
        "405":
          description: 許可されていないメソッド
          schema:
            type: string
      summary: ログイン
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 「Bearer {JWT}」の形式で入力してください
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"